+++ test/threat_policy_test.dart
@@ -32,11 +32,13 @@ void main() {
         Threat.repackaging,
         Threat.trustedInstall,
         Threat.debug,
+        Threat.adbEnabled,
         Threat.devicePasscode,
         Threat.obfuscationIssues,
         Threat.multiInstance,
         Threat.secureHardwareNotAvailable,
         Threat.locationSpoofing,
+        Threat.deviceBinding,
       });
     });